Added Category Entity and Security service

+++ src/main/java/am/devvibes/buyandsell/service/itemForSell/impl/ItemForSellServiceImpl.java
@@ -2,11 +2,13 @@ package am.devvibes.buyandsell.service.itemForSell.impl;
 
 import am.devvibes.buyandsell.dto.itemForSell.ItemForSellRequestDto;
 import am.devvibes.buyandsell.dto.itemForSell.ItemForSellResponseDto;
-import am.devvibes.buyandsell.entity.ItemForSell;
+import am.devvibes.buyandsell.entity.ItemForSellEntity;
 import am.devvibes.buyandsell.exception.NotFoundException;
+import am.devvibes.buyandsell.exception.SomethingWentWrongException;
 import am.devvibes.buyandsell.mapper.ItemForSellMapper;
 import am.devvibes.buyandsell.repository.ItemForSellRepository;
 import am.devvibes.buyandsell.service.itemForSell.ItemForSellService;
+import am.devvibes.buyandsell.service.security.SecurityService;
 import am.devvibes.buyandsell.util.ExceptionConstants;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -20,19 +22,21 @@ public class ItemForSellServiceImpl implements ItemForSellService {
 
 	private final ItemForSellRepository itemForSellRepository;
 	private final ItemForSellMapper itemForSellMapper;
+	private final SecurityService securityService;
 
 	@Override
 	@Transactional
 	public ItemForSellResponseDto saveItemForSell(ItemForSellRequestDto itemForSellRequestDto) {
-		ItemForSell itemForSellEntity = itemForSellMapper.mapDtoToEntity(itemForSellRequestDto);
-		ItemForSell savedItemForSell = itemForSellRepository.save(itemForSellEntity);
+		ItemForSellEntity itemForSellEntity = itemForSellMapper.mapDtoToEntity(itemForSellRequestDto);
+		itemForSellEntity.setUser(securityService.getCurrentUser());
+		ItemForSellEntity savedItemForSell = itemForSellRepository.save(itemForSellEntity);
 		return itemForSellMapper.mapEntityToDto(savedItemForSell);
 	}
 
 	@Override
 	@Transactional
 	public ItemForSellResponseDto findItemForSellById(Long id) {
-		ItemForSell itemForSellEntity = itemForSellRepository.findById(id)
+		ItemForSellEntity itemForSellEntity = itemForSellRepository.findById(id)
 				.orElseThrow(() -> new NotFoundException(ExceptionConstants.ITEM_NOT_FOUND));
 		return itemForSellMapper.mapEntityToDto(itemForSellEntity);
 	}
@@ -40,14 +44,23 @@ public class ItemForSellServiceImpl implements ItemForSellService {
 	@Override
 	@Transactional
 	public List<ItemForSellResponseDto> findAllItemForSells() {
-		List<ItemForSell> itemForSellEntityList = itemForSellRepository.findAll();
+		List<ItemForSellEntity> itemForSellEntityList = itemForSellRepository.findAll();
 		return itemForSellMapper.mapEntityListToDtoList(itemForSellEntityList);
 	}
 
 	@Override
 	@Transactional
 	public void deleteItemForSell(Long id) {
+		ItemForSellEntity itemForSellEntity = itemForSellRepository.findById(id)
+				.orElseThrow(() -> new NotFoundException(ExceptionConstants.ITEM_NOT_FOUND));
+		if (!itemForSellEntity.getUser().getId().equals(securityService.getCurrentUser().getId())) {
+			throw new SomethingWentWrongException(ExceptionConstants.INVALID_ACTION);
+		}
 		itemForSellRepository.deleteById(id);
 	}
+
+	//	private ItemForSellEntity setUser(ItemForSellEntity itemForSellEntity) {
+	//		itemForSellEntity.setUser();
+	//	}
 
 }
